MoviePipeline: Fixed several actions in the Queue UI (such as converting to jobs, or assigning configurations) not supporting multi-select. Previously only toggling shots/jobs supported multi-select, but now converting job types, and assigning job/shot configurations should also respect multi-select.
If you have multiple items selected and you use the drop-down on a non-selected item, only the item for which the drop-down was clicked will be changed (as it was not part of the multi-select, so we're considering it not a multi-select action). If the item who's drop-down you clicked was part of the multi-select then the change will be propagated to same-type jobs/shots (ie: if you have both graph and non-graph jobs selected and you click the dropdown for a graph job and choose a new asset, the multi-selected graph jobs will use that new configuration but no changes will be made to non-graph jobs). When converting between graph and non-graph configurations, jobs that are already of the target type will be unchanged.

Specific scenarios tested:
1) Having a Config-based job selected, a child shot selected, and a second config-based job selected, and A) assigning a Job Configuration via the dropdown, B) clearing a Job Configuration, C) applying and clearing a shot-configuration D) converting to graph jobs
2) The above, but with a Graph-based job selected as part of the mix.
3) Having a Graph based job selected, a child shot selected, and a second graph-based job selected, and A) assigning a Graph Configuration via the dropdown, B) creating a new Graph (for both the job and the sub-shot), C) replacing with preset (converting back to non-graph job)
4) The above, but with a non-graph job selected as part of the mix

#jira UE-218811
#rb chad.williams

[CL 35317562 by matt hoffman in ue5-main branch]