[Backout] - CL35095370
Fixed the missing bitfield initialization and invalidated shaders for good measure.

[FYI] Zousar.Shaker
Original CL Desc
-----------------------------------------------------------------
[Backout] - CL35079483
[FYI] tim.doerries
Original CL Desc
-----------------------------------------------------------------
Added a new FirstPersonOutput node which allows users to pass a FirstPersonInterpolationAlpha value to interpolate between world space and first person space on a per-vertex frequency.
Apart from the expected plumbing required, it was also necessary to add the option for custom output nodes to compile for the previous frame. This is required because this feature allows the material to change the mesh positions, so we need to evaluate for the current and previous frame to get correct motion vectors.
In addition, there is an optional optimization for WPO where materials modifying the mesh positions can be replaced with a default material unless the primitive is tagged as bEvaluateWorldPositionOffset. This optimization seems to assume that WPO is the only source of such modifications, so some mesh processors had to have some additional checks added to keep the original material if it is a first person primitive and using the new node. First person rendering works fine with the default material in the absence of the new node because it is fixed functionality compiled into every material.

Reviewed in 35016692

#rb Sebastien.Hillaire

[CL 35346012 by tim doerries in ue5-main branch]