[ChaosVD] Adding support to record data outside a solver stage.
Initially, CVD only supported recording data within a recorded solver stage, everything else was discarded and an ensure was triggered.

But currently, we also record events that might not fall withing a predefined stage (like disabling a particle or creating a particle). For these cases we added them to the last available stage on load, and discard them if there was none available.

This CL add the possibility to properly create a new solver stage entry during load, if we reach one of these scenarios. This allow us to ensure the data is applied in the correct order, and clearly see where the change of state (if any came from) as it will be shown as an stage named "In Between Stage Data".

#jira UE-214135
#rb Chris.Caulfield

[CL 35389260 by sergio gardeazabal in ue5-main branch]