EditorDataStorageFeatures : Add the TedsSettings module
The TedsSettings module is responsible for mirroring settings data from the ISettingsModule into the editor data storage.  There is 1 row per ISettingsSection, and a pointer to the SettingsObject ( UObject ) for the section is stored in the row using the compatible object API.

Individual settings are properties of the SettingsObject, and it is up to the different users of this row data to first query for the settings object before inspecting or modifying its properties.  This offers the same granularity and mode of operation as the ISettingsSection.

Columns are added for the SettingsContainer, SettingsCategory, and SettingsSection.  These currently contain only the respective FName for each, but could contain more data ( e.g. DisplayName, Description ) or pointers ( e.g. to the ISettingsCategory ) in the future.

A console variable "TEDS.Feature.Settings.Enable" is provided to enable this feature, and it is disabled by default at this time.

#jira UE-220810
#rb ronald.koppers

[CL 35393978 by ross smith in ue5-main branch]