Sequencer: Fixed pre-animated state being cached too late for skel animations
Since pre-animated state was only being tracked and cached inside the Evaluation phase, any changes that removed tracked entities could lead to a temporary state whereby no tracking exists for skeletal animation (ie, when manipulating a skeletal animation track) because pre-animated state restoration occurs at the end of the instantiation phase, _before_ the new entity is evaluated. This was leading to unnecessary UI flicker caused by re-initialization of the anim instance. The fix is to move the pre-animated state tracking into the instantiation phase and ensure this runs before the RestorePreAnimatedState system is allowed to run.

#jira UE-215502
#rb ludovic.chabant, David.Bromberg

[CL 35400704 by andrew rodham in ue5-main branch]