Sequencer: Deprecated SceneComponentBinding components in favor of a more flexible and generic BoundObjectResolver component
This new component allows sections to specify any function ptr to map a boundobject to the object it actually needs to animate. For Transforms this is a USceneComponent, for skeletal animations this is a USkeletalMeshComponent. The methods by which these are resolved varies by track and is now specified through this additional component.

#rb david.bromberg, ludovic.chabant

[CL 35400779 by andrew rodham in ue5-main branch]