Sequencer: Changes to ensure old style spawnables aren't created in UEFN.
First, dragging an asset into Sequencer was allowing old style spawnables in UEFN because of the fallback in CreateGenericBinding and AllowsSpawnableObjects was allowing any spawn tracks/sections. With this change, we only allow old style spawnables if there are no custom binding types (ie. UMG).

Second, shift-drag into the viewport (FSequencer::OnNewActorsDropped) was creating spawnables because it didn't check for supported custom binding types.

AllowsSpawnableObjects now returns true if there are any custom binding types that are UMovieSceneSpawnableBindingBase (this still allows for UFortNPCCharacterMovieSceneSpawnableBinding for UEFN.

#rb Andrew.Rodham, david.bromberg

[CL 35432171 by max chen in ue5-main branch]