Add "(Digital)" to the gamepad right and left trigger gamepad key display names in the editor.
This should hopefully help alleviate any confusion when mapping to these keys, since their value is only going to ever be 0.0 or 1.0. In order to get any values in between (i.e. the exact value of the trigger) you should bind to the gamepad trigger "Axis".

Only make these changes in the editor in case there is any gamepad/UI code displaying the name of the keys that game's don't want to change.

#rnx
#rb jibb.smart

[CL 35451014 by ben hoffman in ue5-main branch]