Added support for TEDS Query Callbacks to be called per chunk on different threads.
Tables in TEDS are broken up in smaller chunks that get processed by calling Query Callbacks for each of them. This typically happens on the calling thread, meaning all chunks are processed in sequence. This updates adds an new option that allows the chunks to be processed in parallel on multiple threads. In some cases this can improve performance, if the processors don't have any internal dependencies on each other.
This change also cleans up some outdated alias uses.

#jira UE-220031
#rb Julien.StJean, logan.buchy

[CL 35455824 by ronald koppers in ue5-main branch]