Universal Search prototype - Core and UI + TEDS Search Provider
This change introduces the core framework and UI for the Universal Search. It doesn't include any search providers and therefore can't search yet. It does however include:
- an input bar that can be docked in the bottom status bar (default) or top main menu.
- configuration options.
- a search providers interface that support async searches.
- a result window with filters, groups and search results.

Adds support to the Universal Search to search through TEDS. Currently only FString, FText and FName are supported. Properties on columns that have the meta data tag "Searchable" will be included in the search. Several columns in TEDS have gotten this meta tag with in this change list. For each combination of group and column a unique query is created so the work can be time sliced better. A miscellaneous query is also created that excludes all other groups so searchable columns that weren't in a group still produce results.

Seperate CRs can be found under 35437674 and 35446965.

#jira UE-220031
#rb logan.buchy, brooke.hubert
#rnx

[CL 35535249 by ronald koppers in ue5-main branch]