Throttle waiting for shared buffers to allow progress
On some platforms, the list of active threads is only modified if a
higher-priority thread wakes up. If no thread in that list ever enters a
wait/stop state and progress is blocked, a deadlock occurs. The yield doesn't
help as it just moves a thread to the back of the queue. Everything gets a
slice of the CPU pie if we drop to the lowest priority thread (even those that
don't yield as the scheduler in question timeslices lowest priority threads).

#ushell-cherrypick of 35588805 by Martin.Ridgers
#rb johan.berg
#rnx

[CL 35592116 by martin ridgers in ue5-main branch]