Generate VarGet for PointerToReference and Freeze for ReferenceToValue
This paves the way for the same bytecode to support both pure-Verse deep mutability using VVars, and native deep mutability using VNativeRef proxy objects (coming in a future patch). Pure Verse now consistently represents a "pointer" with a VVar and a "reference" with a melted VCell; native interop will represent both a "pointer" and a "reference" as a VNativeRef. In the bytecode for an assignment, the sequence of VarGets, LoadFields, and Calls (without intervening Freezes) followed by a VarSet, StoreField, or CallSet, will thus function as the union of the operations necessary for both representations.

The semantic analyzer transforms some assignments into multiple statements, with intermediate references stored in locals. (This supports parameter_collection notifications and custom class var accessors.) These transformations did not always use CExprPointerToReference and CExprReferenceToValue in a way consistent with user-written statements, so they have been adjusted to match. They also now use immutable locals when possible for efficiency in the new VM. Finally, because the new VM already has distinguished LoadField and StoreField instructions, the transformation to support custom class var accessors is disabled there, further simplifying the generated bytecode.

With these changes, CompileAssignmentImpl only needs to consider the top level AST node on the assignment LHS, because it must become a VarSet, StoreField, or CallSet. Child nodes can be compiled with CompileExpression, which will produce a "reference" usable with these instructions.

#rb saam.barati

[CL 35658837 by russell johnston in ue5-main branch]