ProceduralDaySequence: Create a base struct that is an abstraction of a procedural day sequence
This removes AProceduralDaySequenceActor and ASunAngleDaySequenceActor. The former was removed because it was made obsolete, while the latter was removed because it was refactored into a procedural sequence. ASunPositionDaySequenceActor remains because it does more than just procedural sequence generation.

We add the notion of a "procedural sequence" which can be configured inside of a collection asset. Any DSA can use a procedural sequence.

#rb lonnie.li

#virtualized

[CL 35666849 by nickolas drake in ue5-main branch]