Sequencer: Fix duplicate binding not working in UEFN.
When pasting a copied binding, if the pasted binding has actors that are already bound, it prompts the user whether to duplicate the bound actor. The search for whether the actors that are already bound uses GetPathName which didn't supply an outer. That didn't match up with when the bindings are copied, which calls GetPathName with an outer. This fix just matches up those calls.

#rb david.bromberg

[CL 35739314 by max chen in ue5-main branch]