Fix incorrect eye / hair material conversion with Substrate.
When using SubstrateMaterialAttribute or SubstrateShadingModel, with a shading model driven by expression, the resulting packing was incorrect for hair and eye material. This is because the substrate operator was converted as a BSDF operator (vs. an eye / hair operator), and later the material was flagged as 'single' instead of 'complex'. The substrate export code was thus missing the complex header packing causing eye/hair data to be incorrect.

This CL add hair & eye BSDF feature, and refactor the feature tracking through a bitfield for simplicity.

#rb sebastien.hillaire

[CL 35770662 by charles derousiers in ue5-main branch]