[Backout] - CL35773939
Partial back-out of this CL (MessageNodes still allow hookups to redirected static functions so as not to break potentially submitted content).

[FYI] jodon.karlik
Original CL Desc
-----------------------------------------------------------------
K2Nodes now allow a CoreRedirect from the "self" pin if (and only if) the new function changes the visibility of the "self" pin.  This allows you to redirect member functions to non-member static functions (which will hide the self pin).  For example:

If the original call was a member function Foo.Bar(123) you could redirect that to a static function Bar(Foo, 123).

Syntax in BaseEngine.ini [[CoreRedirects] looks something like:
; Redirect the Member Function to a Static Function
+FunctionRedirects=(OldName="/Script/PluginName.Foo.Bar", NewName="FooBlueprintLibrary.Bar")
; Redirect the Self Pin to a Parameter Pin named "InFoo"
+PropertyRedirects=(OldName="FooBlueprintLibrary.Bar.Self", NewName="InFoo")

#rb ben.zeigler

[CL 35787966 by jodon karlik in ue5-main branch]