Crunched FTextKey down to a single int32
FTextKey used to be a pointer and embedded hash (when UE_TEXTKEY_STORE_EMBEDDED_HASH=1), where the pointer was a reference to an entry in an internal table.
This change updates FTextKey to just store the index to the entry in the internal table, and then look-up the data when needed. This look-up is fast enough that UE_TEXTKEY_STORE_EMBEDDED_HASH can now be disabled by default.

These changes take FTextKey from 12 bytes to 4 bytes, and FTextId from 24 bytes to 8 bytes. This can save a few MB of memory in the display string live table, plus makes anything else storing a FTextKey/FTextId (like the FText history types) smaller too.

#rb Leon.Huang

[CL 35867334 by jamie dale in ue5-main branch]