DaySequence: Rework static time
Our current static time solution has a number of issues
- It cannot be blended
- Modifiers can disable static time at unexpected times (stomping static time set by a different modifier)
- It can't be changed without reenabling a modifier

I am introducing the notion of the StaticTimeManager and StaticTimeContributors. The manager has an API for adding and removing contributors and getting the current static time. Contributors just provide the manager with some callbacks for their desired static time and their weights and the manager handles resolving the effective static time.

#rb lonnie.li

#changelist validated
#virtualized

[CL 35885531 by nickolas drake in ue5-main branch]