PathTracing: Implement new debug visualization mode
This mode is similar to the raytracing debug visualization, but using the path tracer's payload. In primary ray mode, a basic light loop is run without sample accumulation which can be used to test performance of the path tracer's light sampling. This mode also allows TSR to run, which can act as a rudimentary denoiser of the lighting signal if enough samples are used.

Remove light grid and decal grid visualization cvars from the main path tracer as these are now well covered by the new debug mode.

#rnx

[CL 35932255 by chris kulla in ue5-main branch]