Tests: Streaming: Fixed async loading tests failing because of some of issues with assets (System.Engine.Loading.ThreadSafeAsyncLoadingTest, System.Engine.Loading.UnattendedAsyncLoadingTest)
Tese streaming tests take the first 5000 assets from AssetRegistry and try to stream them. However, some of the assets are either problematic or not expected to be loaded outside of the editor, which leads to loading errors failing the tests.

Some of the problems I've found include:
   - ARUtilities, OpenXR: Trying to load assets from plugins which are not enabled.
   - WorldPartition: HLODLayer objects from the engine referencing HLODBuilderMeshMergeSettings objects implemented on the editor side (these assets are probably not meant to be loaded outside of the editor)
   - Harmonix, MovieRenderPipeline: issues that I haven't been able to identify, possibly broken assets.

Resolved by limiting which assets we're trying to use in the tests. It doesn't reduce the generality of the tests, as they are meant to test our streaming logic, not validate the correctness of assets. This change is more of a workaround, though.

#rb David.Harvey

[CL 35932829 by wojciech krywult in ue5-main branch]