#JIRA UE-223182 Curve Table Filtering;
Improving the curve table uasset filtering so we get more specific matches to the text queries. For example, typing "location" will filter all text with location in it but then typing "location.x" used to filter everything with location and x in it, now it'll filter things with location.x in it. So instead of filtering strings with x in it like "max", it now keeps both of the terms in consideration by finding x and then searching the hierarchy for location as well.

-- Code --
CurveTableEditor:
FCurveTableEditorItem::PassesFilter() - Updating the string search to not just be a Match() but to find each token in the string's hierarchy. Written by Andrew Rodham.

[REVIEW] [at]andrew.rodham
#tests Editor - opened a curve table uasset and tried filtering by different variations of terms.
#rb Andrew.Rodham

[CL 35972920 by aaron eady in ue5-main branch]