TedsSettingsManager : In UpdateSettingsCategory, always update section rows as their settings object may have changed.
Store an index for each row using a hash of the ContainerName + CategoryName + SectionName so that we can find rows even if they are pending or if their settings object has changed.

This fixes an issue where the same Container / Category / Section is registered twice with 2 different settings objects.  In this case, the end result should be 1 row in TEDS with the settings object from the second call to RegisterSettings.  Added a test case for regression testing.

#jira UE-220810
#rb aditya.ravichandran

[CL 35986403 by ross smith in ue5-main branch]