MPCook: CookPackageSplitter: Add RequiresCachedCookedPlatformDataBeforeSplit function for ICookPackageSplitter classes, and add a save substate to support it by waiting for IsCachedCookedPlatformDataLoaded to return true before calling ShouldSplit.
Refactor ESaveState FGenerationHelper to instead be ESaveSubState on FPackageData. This reduces the complexity of maintaining two state machines for the save state. The first state machine was the de facto state machine created out of bools on the FPackageData, and the second was the ESaveState state machine on the FGenerationHelper. They are now merged into the single ESaveSubState state machine.
#jira UE-222507
#rnx
#rb Zousar.Shaker

[CL 36012081 by matt peters in ue5-main branch]