Add engine support for primitive instance data that resides only on the GPU, along with hook points for populating this instance data.
Saves CPU time and memory for applications where CPU-side rendering systems are not required. Current list

* Affecting Lumen lighting
* Affecting Distance Field lighting
* Ray tracing
* Baked lighting
* Other CPU systems such as physics collision, navigation not supported by design

Also, non-Nanite meshes and instance hierarchy culling are not supported yet.

This feature is not exposed to the current primitive component hierarchy. It will be opted in via code in a new procedural ISM compoonent in a subsequent commit.

#rb Jeremy.Moore, Ola.Olsson
#jira UE-217681
#rnx

[CL 36033060 by huw bowles in ue5-main branch]