Implement proper visibility of new viewport toolbar separators
Set the visibility of separators to a delegate that depends on the visibility of previously and upcoming raised entries. This allows separators to only be visible when they are surrounded by raised entries and avoids situations where dynamically raised entries are hidden, leaving separators as the first or last elements in a toolbar.

This is needed for the new viewport toolbar. We raise entries from toolbar submenus (combo buttons) so they appear directly in the top-level viewport toolbar. The condition for raising entries can be dynamically driven by a TAttribute<bool> delegate. We want to add separators between raised entries if said entries either live in separate sections or have an explicit separator between them. Because entries are dynamically visible but we still want to show separators when the entries are visible, the visibility of the separators themselves has to be dynamic which this change supports.

#jira UE-212285
#rb ross.smith

[CL 36033530 by sebastian arleryd in ue5-main branch]